+++ tests/testthat/test-LinkedMatrix.R
@@ -1,15 +1,11 @@
-# Prepare dummy data
-dummy <- matrix(data = seq_len(16), nrow = 4, ncol = 4)
-rownames(dummy) <- paste0("row_", seq_len(nrow(dummy)))
-colnames(dummy) <- paste0("col_", seq_len(ncol(dummy)))
+n <- 4L
+p <- 4L
+dimnames <- list(
+    paste0("row_", seq_len(n)),
+    paste0("col_", seq_len(p))
+)
 
-createLinkedMatrix <- function(class, nNodes) {
-    linkedBy <- ifelse(class == "ColumnLinkedMatrix", "columns", "rows")
-    linkedMatrix <- LinkedMatrix(nrow = nrow(dummy), ncol = ncol(dummy), nNodes = nNodes, linkedBy = linkedBy, nodeInitializer = "matrixNodeInitializer")
-    linkedMatrix[] <- dummy
-    dimnames(linkedMatrix) <- dimnames(dummy)
-    return(linkedMatrix)
-}
+dummy <- createMatrix(n, p, dimnames)
 
 for (class in c("ColumnLinkedMatrix", "RowLinkedMatrix")) {
 
@@ -21,12 +17,12 @@ for (class in c("ColumnLinkedMatrix", "RowLinkedMatrix")) {
 
         for (nNodes in c(1, 2)) {
 
-            linkedMatrix <- LinkedMatrix(nrow = nrow(dummy), ncol = ncol(dummy), nNodes = nNodes, linkedBy = linkedBy, nodeInitializer = "matrixNodeInitializer")
+            linkedMatrix <- LinkedMatrix(nrow = n, ncol = p, nNodes = nNodes, linkedBy = linkedBy, nodeInitializer = "matrixNodeInitializer")
             expect_equal(nNodes(linkedMatrix), nNodes)
             expect_is(linkedMatrix[[1]], "matrix")
 
             if (requireNamespace("ff", quietly = TRUE)) {
-                linkedMatrix <- LinkedMatrix(nrow = nrow(dummy), ncol = ncol(dummy), nNodes = nNodes, linkedBy = linkedBy, nodeInitializer = "ffNodeInitializer", vmode = "integer")
+                linkedMatrix <- LinkedMatrix(nrow = n, ncol = p, nNodes = nNodes, linkedBy = linkedBy, nodeInitializer = "ffNodeInitializer", vmode = "integer")
                 expect_equal(nNodes(linkedMatrix), nNodes)
                 expect_is(linkedMatrix[[1]], "ff_matrix")
             }
@@ -50,7 +46,7 @@ for (class in c("ColumnLinkedMatrix", "RowLinkedMatrix")) {
         expect_equal(dim(linkedMatrix), c(1, 1))
 
         # Single LinkedMatrix input
-        linkedMatrix <- new(class, createLinkedMatrix(class, 2))
+        linkedMatrix <- new(class, createLinkedMatrix(n, p, dimnames, class, 2))
         expect_equal(nNodes(linkedMatrix), 1)
         expect_equal(dim(linkedMatrix), dim(dummy))
 
@@ -64,12 +60,12 @@ for (class in c("ColumnLinkedMatrix", "RowLinkedMatrix")) {
         }
 
         # Multiple LinkedMatrix inputs of same order
-        linkedMatrix <- new(class, createLinkedMatrix(class, 2), createLinkedMatrix(class, 2))
+        linkedMatrix <- new(class, createLinkedMatrix(n, p, dimnames, class, 2), createLinkedMatrix(n, p, dimnames, class, 2))
         expect_equal(nNodes(linkedMatrix), 2)
         if (class == "ColumnLinkedMatrix") {
-            expect_equal(dim(linkedMatrix), c(nrow(dummy), ncol(dummy) * 2))
+            expect_equal(dim(linkedMatrix), c(n, p * 2))
         } else {
-            expect_equal(dim(linkedMatrix), c(ncol(dummy) * 2, nrow(dummy)))
+            expect_equal(dim(linkedMatrix), c(p * 2, n))
         }
 
         # Multiple conformable matrix inputs of different order
@@ -116,95 +112,12 @@ for (class in c("ColumnLinkedMatrix", "RowLinkedMatrix")) {
         }
     })
 
-    for (nNodes in seq_len(ifelse(class == "ColumnLinkedMatrix", ncol(dummy), nrow(dummy)))) {
+    for (nNodes in seq_len(ifelse(class == "ColumnLinkedMatrix", p, n))) {
 
         context(paste0(class, " with ", nNodes, " nodes"))
 
         # Prepare LinkedMatrix object
-        linkedMatrix <- createLinkedMatrix(class, nNodes)
-
-        test_that("replacement", {
-
-            # Generate new dummy for replacement
-            replacement <- matrix(data = seq_len(16) * 10, nrow = 4, ncol = 4)
-            rownames(replacement) <- paste0("row_", seq_len(nrow(replacement)))
-            colnames(replacement) <- paste0("col_", seq_len(ncol(replacement)))
-            comparison <- dummy
-
-            idx2 <- expand.grid(seq_len(nrow(dummy)), seq_len(ncol(dummy)))
-
-            testAndRestore <- function(info) {
-                expect_equal(linkedMatrix[], comparison, info = info)
-                linkedMatrix <- createLinkedMatrix(class, nNodes)
-                assign("linkedMatrix", linkedMatrix, parent.frame())
-                assign("comparison", dummy, parent.frame())
-            }
-
-            linkedMatrix[] <- replacement
-            comparison[] <- replacement
-            testAndRestore("[]")
-
-            for (i in seq_len(nrow(dummy))) {
-
-                linkedMatrix[i, ] <- replacement[i, ]
-                comparison[i, ] <- replacement[i, ]
-                testAndRestore(paste0("[", i, ", ]"))
-
-                linkedMatrix[i, ] <- NA
-                comparison[i, ] <- NA
-                testAndRestore(paste0("[", i, ", ] <- NA"))
-
-            }
-
-            for (i in seq_len(ncol(dummy))) {
-
-                linkedMatrix[, i] <- replacement[, i]
-                comparison[, i] <- replacement[, i]
-                testAndRestore(paste0("[, ", i, "]"))
-
-                linkedMatrix[, i] <- NA
-                comparison[, i] <- NA
-                testAndRestore(paste0("[, ", i, "] <- NA"))
-
-            }
-
-            for (i in seq_len(nrow(idx2))) {
-
-                linkedMatrix[idx2[i, 1], idx2[i, 2]] <- replacement[idx2[i, 1], idx2[i, 2]]
-                comparison[idx2[i, 1], idx2[i, 2]] <- replacement[idx2[i, 1], idx2[i, 2]]
-                testAndRestore(paste0("[", idx2[i, 1], ", ", idx2[i, 2], "]"))
-
-                linkedMatrix[idx2[i, 1]:idx2[i, 2], ] <- replacement[idx2[i, 1]:idx2[i, 2], ]
-                comparison[idx2[i, 1]:idx2[i, 2], ] <- replacement[idx2[i, 1]:idx2[i, 2], ]
-                testAndRestore(paste0("[", idx2[i, 1], ":", idx2[i, 2], ", ]"))
-
-                linkedMatrix[, idx2[i, 1]:idx2[i, 2]] <- replacement[, idx2[i, 1]:idx2[i, 2]]
-                comparison[, idx2[i, 1]:idx2[i, 2]] <- replacement[, idx2[i, 1]:idx2[i, 2]]
-                testAndRestore(paste0("[, ", idx2[i, 1], ":", idx2[i, 2], "]"))
-
-                linkedMatrix[idx2[i, 1]:idx2[i, 2], idx2[i, 1]:idx2[i, 2]] <- replacement[idx2[i, 1]:idx2[i, 2], idx2[i, 1]:idx2[i, 2]]
-                comparison[idx2[i, 1]:idx2[i, 2], idx2[i, 1]:idx2[i, 2]] <- replacement[idx2[i, 1]:idx2[i, 2], idx2[i, 1]:idx2[i, 2]]
-                testAndRestore(paste0("[", idx2[i, 1], ":", idx2[i, 2], ", ", idx2[i, 1], ":", idx2[i, 2], "]"))
-
-                linkedMatrix[c(idx2[i, 1], idx2[i, 2]), ] <- replacement[c(idx2[i, 1], idx2[i, 2]), ]
-                comparison[c(idx2[i, 1], idx2[i, 2]), ] <- replacement[c(idx2[i, 1], idx2[i, 2]), ]
-                testAndRestore(paste0("[c(", idx2[i, 1], ", ", idx2[i, 2], "), ]"))
-
-                linkedMatrix[, c(idx2[i, 1], idx2[i, 2])] <- replacement[, c(idx2[i, 1], idx2[i, 2])]
-                comparison[, c(idx2[i, 1], idx2[i, 2])] <- replacement[, c(idx2[i, 1], idx2[i, 2])]
-                testAndRestore(paste0("[, c(", idx2[i, 1], ", ", idx2[i, 2], ")]"))
-
-                linkedMatrix[c(idx2[i, 1], idx2[i, 2]), c(idx2[i, 1], idx2[i, 2])] <- replacement[c(idx2[i, 1], idx2[i, 2]), c(idx2[i, 1], idx2[i, 2])]
-                comparison[c(idx2[i, 1], idx2[i, 2]), c(idx2[i, 1], idx2[i, 2])] <- replacement[c(idx2[i, 1], idx2[i, 2]), c(idx2[i, 1], idx2[i, 2])]
-                testAndRestore(paste0("[c(", idx2[i, 1], ", ", idx2[i, 2], "), c(", idx2[i, 1], ", ", idx2[i, 2], ")]"))
-
-                linkedMatrix[idx2[i, 1]:idx2[i, 2], idx2[i, 1]:idx2[i, 2]] <- NA
-                comparison[idx2[i, 1]:idx2[i, 2], idx2[i, 1]:idx2[i, 2]] <- NA
-                testAndRestore(paste0("[", idx2[i, 1], ", ", idx2[i, 2], "] <- NA"))
-
-            }
-
-        })
+        linkedMatrix <- createLinkedMatrix(n, p, dimnames, class, nNodes)
 
         test_that("dim", {
             expect_equal(dim(linkedMatrix), dim(dummy))
@@ -223,7 +136,7 @@ for (class in c("ColumnLinkedMatrix", "RowLinkedMatrix")) {
             if (class == "RowLinkedMatrix") {
 
                 boundLinkedMatrix <- rbind(linkedMatrix, linkedMatrix)
-                expect_equal(dim(boundLinkedMatrix), c(nrow(dummy) * 2, ncol(dummy)))
+                expect_equal(dim(boundLinkedMatrix), c(n * 2, p))
                 expect_equal(nNodes(boundLinkedMatrix), nNodes * 2)
 
                 expect_error(cbind(linkedMatrix, linkedMatrix))
@@ -231,7 +144,7 @@ for (class in c("ColumnLinkedMatrix", "RowLinkedMatrix")) {
             } else {
 
                 boundLinkedMatrix <- cbind(linkedMatrix, linkedMatrix)
-                expect_equal(dim(boundLinkedMatrix), c(nrow(dummy), ncol(dummy) * 2))
+                expect_equal(dim(boundLinkedMatrix), c(n, p * 2))
                 expect_equal(nNodes(boundLinkedMatrix), nNodes * 2)
 
                 expect_error(rbind(linkedMatrix, linkedMatrix))