+++ NEWS.md
@@ -1,6 +1,7 @@
 # LinkedMatrix 1.3.0.9000
 
-* Add [crochet](https://cran.r-project.org/package=crochet) subsetting support.
+* Add [crochet](https://cran.r-project.org/package=crochet) subsetting and
+  replacement support.
 * Add generic `as.ColumnLinkedMatrix` and `as.RowLinkedMatrix` for easy
   creation of `LinkedMatrix` objects from `list`s of matrix-like objects
   without `do.call`.